+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Gunpowder Market 1.0.4
+
+- Update to 1.17
+- Remove caching
+- Use proper column types
+
+---
 # Gunpowder Market 1.0.3
 
 - Remove debug code that caused issues.